# docs/swagger.yaml
...
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.SearchObject:
    properties:
      accepted:
        type: boolean
      answer_count:
        type: integer
      created_at:
        type: integer
      excerpt:
        type: string
      id:
        type: string
      question_id:
        type: string
      status:
        description: Status
        type: string
      tags:
        description: tags
        items:
          $ref: '#/definitions/schema.TagResp'
        type: array
      title:
        type: string
      url_title:
        type: string
      user_info:
        allOf:
        - $ref: '#/definitions/schema.SearchObjectUser'
        description: user info
      vote_count:
        type: integer
    type: object
  schema.SearchObjectUser:
    properties:
      display_name:
        type: string
      id:
        type: string
      rank:
        type: integer
      status:
        type: string
      username:
        type: string
    type: object
  schema.SearchResp:
    properties:
      count:
        type: integer
      list:
        description: search response
        items:
          $ref: '#/definitions/schema.SearchResult'
        type: array
    type: object
  schema.SearchResult:
    properties:
      object:
        allOf:
        - $ref: '#/definitions/schema.SearchObject'
        description: this object
      object_type:
        description: object_type
        type: string
    type: object
...
  schema.TagResp:
    properties:
      display_name:
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/search:
    get:
      description: search object
      parameters:
      - description: query string, can be empty when tag_groups is given
        in: query
        name: q
        type: string
      - description: order
        enum:
        - newest
        - active
        - score
        - relevance
        in: query
        name: order
        required: true
        type: string
      - collectionFormat: multi
        description: tag groups
        in: query
        items:
          type: string
        name: tag_groups
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.SearchResp'
              type: object
      security:
      - ApiKeyAuth: []
      summary: search object
      tags:
      - Search
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"